Make Test Deterministic
This resolves a test that would occasionally fail. Ultimately, the
number of testimonials to display relies on a coin flip to determine if
it should show no testimonials or one, if no other conditions are
satisfied.

Previously, to ensure that we would see all possible output, the test
was run twice, and we checked to see if we received both 0 and 1 as
output. However, because this is testing ruby randomization,
statistically, we're not guaranteed to always see both values, no matter
how many iterations we increase it to.

As such, instead we are running it once, and check to see that the value
is either 0 or 1. This does not exercise/show each possible scenario as
output, but it should not cause the test suite to incorrectly fail,
depending on the odds of a coin flip coming up both heads and tails in
successive flips.