Explicitly state regexes in relevant examples
Rather than extracting the regex defining a company or user access key
in the tests that check to see the keys determined look as they should,
this defines the regex multiple times within the examples that need it.

This makes it more clear within the relevant example to make it obvious
what the expectation is that it should be matching. While this does mean
that the same code is written multiple times, and it may need to be
updated should it change over time in different places, it adds
explanatory value and doesn't require a person to navigate around a file
to find the value.