Ask the accessor to generate its access key
With the logic about how to generate an access key separated, this now
changes the public API of the `AccessKeyGenerator` to accept an instance
of the class that needs an access key generated and calls the
appropriate method to do so.

The tests have been modified to pass. This should then, with adequate
test coverage in place, signal on a test run every place that the usage
of this class needs to change, because the signature has changed
significantly. The work to change the calling of this method can be done
at this point, and we can stop here, or we can use this as a signal for
every place where we can replace this class entirely by instead calling
the model's `generate_access_key` method directly.